refactor: 맵뷰 역할별로 MapViews 폴더에 정리
