+++ TrackUs/TrackUs/Sources/Running/View/RunningResultView.swift
@@ -89,7 +89,7 @@ extension RunningResultView {
     
     var body: some View {
         VStack {
-            RouteMapView(coordinates: trackingViewModel.coordinates)
+            PathPreviewMap(coordinates: trackingViewModel.coordinates)
             
             VStack {
                 VStack(spacing: 20) {
